fix: don't build cdylib when testing
On macOS a cdylib that exposes Python API symbols must either link to a Python framework/lib or be built with -undefined dynamic_lookup. Because those flags aren’t present in the 'cargo test', the cdylib link step sees unresolved Py* symbols and the build fails.

This change pushes the cdylib to a feature flag. maturin knows to enable the pyo3/extension-module